fix: url() check should include subdomain()
When .url was created we only had path gateways.  When .subdomain was
added, we did not update .url to test for subdomain gateways, which in
the long run will confuse people and feels like a bug.

Let's fix this: .url() will now check for both subdomain and path gateways

https://github.com/ipfs/is-ipfs/pull/32#discussion_r396161665

BREAKING CHANGE: .url(url) now returns true if .subdomain(url) is true

License: MIT
Signed-off-by: Marcin Rataj <lidel@lidel.org>